Update pal from commit 48ad0c6e
* Update RDF
* Update DevDriver to v23.11.22
* Update TraceSession::CollectTrace Result codes based on trace state
* Update platform settings to devdriver settings
* Update core settings to devdriver settings
* Update amdgpu_drm.h
* Make targets for settings generation.
* Correct CreateGraphicsPipeline in DeviceDecorator derived classes
* Relax the PipelineAbiProcessor and handle unnamed symbols
* Do not cache symbols with type 'section'
* Initialize GpuPerfExpTraceSource
* Add setting to control DCC MIN_COMPRESSED_BLOCK_SIZE
* Add ConvertWcharStringToUtf8
* Add a new function to Util::Thread class to assign a name to a thread
* Add the support of setting VGT_STRMOUT_CONFIG.EN_PRIMS_NEEDED_CNT
* Add missing L2 flush after early phase gfx blit.
* CodeGen: Add default value for string size
* Support for SQTT & SPM UberTrace chunks
* Move settings source files from json to yaml
* Prepare change for later WaitSync change
* Totally remove the gfx9 P2P BAR workaround
* Dump PM4 in Exe named directory
* Conditional TraceSource activation
* Suppress warning from [[unlikely]] on Clang 10/11 and gcc
* VK_EXT_device_fault - Driver Implementation - Linux
* Couple of barrier code cleanup
* Generic Gang-Submit Cleanups
* Change indirect dispatch/draw interfaces to take GPUVAs
* Remove some deprecated references to addr1 defines that are no longer needed
* Remove code referencing pre-Navi AddrLib enums
* Remove CMake overrides for Gfx9 address-lib
* Remove Validation v nonValidation Distinction from the independent layer
* Remove Gfx9 classes and formats.
* Remove CmdDispatchDynamic support
* Navi3x: set sxblendopt to 0 for alphaToCoverage case
* SX Blend optimization fix with alpha to coverage state
* Fix double-free causing test stop during extended 3dmark testing
* Fix hash size comments
* Fix PAL decorator class missing variables like stackSizeInBytes
* Fix various aspects of app profile-related string handling
* Fix compiler warnings [NFC]
* Fix hangs and crashes in few Vulkan applications
* Fix AddrLib version check
* Fix incorrect application of ppPbbPBBBreakBatchDifferenceWithPrimLimit_FpovLimit_DeallocLimit_A_ WAR
* Fix incorrect usage of sizeof()
* Fix typo PAL_CLIENT_INTERFACE_MAJOR_VERSION guard for CmdDrawIndexedIndirectMulti.
* Fix the failures for dynamicDualSourceBlend also include fast link library cases
* Fix incorrect function name in UberTraceService
* Fix for missing SQTT detailed instruction tokes for some Navi3 GPUs
* Fix Navi3 board power throttle during long time aging

+++ shared/devdriver/shared/ddSettings/base/src/ddSettingsBase.cpp
@@ -254,7 +254,7 @@ DD_RESULT SettingsBase::ApplyUserOverrideImpl(
     DDSettingsValueRef valueRef = {};
     valueRef.hash = useroverride.nameHash;
     valueRef.type = useroverride.type;
-    valueRef.size = useroverride.size;
+    valueRef.size = static_cast<uint16_t>(useroverride.size);
 
     switch (useroverride.type)
     {
